% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addData.R
\name{addData}
\alias{addData}
\title{A function that allows a user to append data to
either the \code{edges} or \code{persons} of a
\code{docNetwork} object and rebiulds the graph.}
\usage{
addData(dnet, data, edges = T, persons = F)
}
\arguments{
\item{dnet}{The \code{docNetwork} object to modify.}

\item{data}{The data to add to the \code{docNetwork}
object.}

\item{edges}{A logical that tells whether the user is
modifying the edges list or not. Default
is True.  Accepts a \code{data.frame},
a character vector of information, or
a path to a file.}

\item{persons}{A logical that tells whether the user is
                modifying the persons list or not. Default
                is False. Accepts a \code{data.frame},
                a character vector of information, or
                a path to a file.}
}
\value{
dnet    The modified \code{docNetwork} object.
}
\description{
A function that allows a user to append data to
either the \code{edges} or \code{persons} of a
\code{docNetwork} object and rebiulds the graph.
}
\examples{
dnet = addData(dnet, data)
dnet = addData(dnet, data, edges=F, persons=T)
dnet = addData(dnet, "/path/to/file.csv")
}

