% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildAuths.R
\name{buildAuths}
\alias{buildAuths}
\title{Creates a data frame of names and their associated TCP numbers}
\usage{
buildAuths(auths, tcp)
}
\arguments{
\item{auths}{A string of \code{;} delimeted author names to be separated.}

\item{tcp}{The \code{TCP} number associated with the authors.}
}
\value{
df   A \code{data.frame} of the authors for a partuclar tcp number
             of the form name, tcp.
}
\description{
Creates a data frame of names and their associated TCP numbers
}
\examples{
authors = buildAuths(tcp$Author[which(tcp$TCP == "A0005")], "A0005")
}

