% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGraph.R
\name{buildGraph}
\alias{buildGraph}
\title{Build the graph for a \code{docNetwork} object.}
\usage{
buildGraph(dnet)
}
\arguments{
\item{dnet}{A docNetwork object}
}
\value{
g An igraph object.
}
\description{
Converts a \code{docNetwork} object into an \code{igraph} object. This step
}
\examples{

g = buildGraph(dnet, draw=T)
g = buildGraph(dnet)

}

