% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityIndex.R
\name{communityIndex}
\alias{communityIndex}
\title{Get index of documents from a community}
\usage{
communityIndex(dnet, community = NULL, method = "internal")
}
\arguments{
\item{dnet}{The \code{docNetwork} object to modify.}

\item{community}{Integer. The membership number of the community you want data about.}

\item{method}{Either "all" or "internal". If "all", results will include all
books attributed to all members of the community. If "internal",
results will include only books that are attributed to two or more
members of the community.}
}
\value{
df    A subset of \code{dnet}'s \code{index} for all nodes in a community.
}
\description{
A function that allows a user to extract bibliographic data from
communities within a larger network graph. Returns a subset of
the index that shows which documents make up a selected community.
}
\examples{
data(dnet)
df = communityIndex(dnet, community = 3, method = "all")


}

