% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communitySubnetwork.R
\name{communitySubnetwork}
\alias{communitySubnetwork}
\title{Subset a network by community}
\usage{
communitySubnetwork(dnet, community, method = "internal")
}
\arguments{
\item{dnet}{The \code{docNetwork} object to modify.}

\item{community}{Integer. The membership number of the community you want data about.}

\item{method}{Either "all" or "internal". If "all", results will include all
books attributed to all members of the community. If "internal",
results will include only books that are attributed to two or more
members of the community.}
}
\value{
dnet    A subsetted \code{docNetwork} for all nodes in a community.
}
\description{
A function that allows a user to extract a \code{docNetwork} from a larger
\code{docNetwork} .
}
\examples{
data(dnet)
df = communitySubnetwork(dnet, community = 3, method = "all")


}

