% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCommunities.R
\name{detectCommunities}
\alias{detectCommunities}
\title{Find communities in a graph}
\usage{
detectCommunities(dnet, type = "walktrap", view = T)
}
\arguments{
\item{dnet}{A document network}

\item{type}{The algorithm to be used: options include "walktrap" and "betweenness".}

\item{view}{A logical that indicates whether you want to view the communities
after detecting them or not. Default is True.}
}
\value{
dnet A \code{dnet} object with a \code{communities} object added into the slot
             that holds the community data detected.
}
\description{
Find communities in a graph
}
\section{Note}{

This function includes only the "walktrap" and "edge.betweenness" options. For a
wider selection of community detection algorithms, consult the \code{igraph} package
directly. Communities identified using alternative methods can be added to a
\code{dnet} object by assigning them to the \code{dnet@communities} slot manually.
\strong{If you want options added, email the developer for requests.}
}
\examples{
communities = detectCommunities(dnet, type="edge.betweenness")
dnet = detectCommunities(dnet, view=F)

}

