% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docNetwork.R
\docType{class}
\name{docNetwork-class}
\alias{docNetwork}
\alias{docNetwork-class}
\title{An S4 class that represents the network of a collection of
 documents.  This object stores the meta-data for the
 social/print network of a corpus.}
\description{
An S4 class that represents the network of a collection of
 documents.  This object stores the meta-data for the
 social/print network of a corpus.
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string that holds the full path to the corpus'
files.}

\item{\code{index}}{A \code{data.frame} with the corpus' metadata.}

\item{\code{edges}}{A \code{data.frame} that stores the connections
between documents and people. In the format
\code{name},\code{TCP}.  Can include more fields.}

\item{\code{persons}}{A \code{data.frame} that stores the authoritative
names and roles for each person in the network.
\code{ID}, \code{Role}.}

\item{\code{graph}}{An \code{igraph} object that holds the graph
of the corpus that is derived from \code{edges}.}

\item{\code{communities}}{A \code{communities} object that stores the
community data derived from \code{graph}.}
}}

