% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drawGraph.R
\name{drawGraph}
\alias{drawGraph}
\title{Draw a social network graph}
\usage{
drawGraph(g, comms = NULL, sub = F, by = 0, color = "blue",
  labelSize = 0.75, nodeSize = 5)
}
\arguments{
\item{sub}{Logical. If true, only a subset of the graph will be printed.}

\item{by}{A numeric or character vector. If only a subset of the graph will
         be drawn, \code{by} controls which part of the graph will be chosen
         for plotting. This can either take the form of an integer (in which
         case it represents the number of a sub-community within the graph)
         or it can be a vector of TCP ids (in which case it selects only a
         subset of the document collection for graphing).}

\item{color}{A string that defines the color of the nodes in the graph.}

\item{labelSize}{A numeric value that sets the size of the labels for the nodes.}

\item{nodeSize}{A numeric value that sets the size of the nodes.}

\item{dnet}{A docNetwork object}
}
\description{
Creates a two-dimensional plot of a social-network graph.
}
\section{How to use it}{

\code{drawGraph} offers only a few options regarding layout. It's designed
to generate easy-to-read results quickly. More advanced plotting
options are available directly through \code{igraph}.
}

