% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportPersonography.R
\name{exportPersonography}
\alias{exportPersonography}
\title{Exports a personography from graph data.  The data is both returned
and written to a file with \code{filename}.}
\usage{
exportPersonography(g, filename = "personography.csv", dnet = NULL)
}
\arguments{
\item{g}{An igraph object.  The graph that the data should be extracted from.}

\item{filename}{A string that gives the name of the file to store the
personography in.  The file will be saved relative to
the \code{directory} stored in your \code{docNetwork}
object. If you do not pass a \code{docNetwork} object
\code{filename} must be the full path for the file.}

\item{dnet}{An optional reference to your \code{docNetwork} object
                 from which the \code{directory} to store \code{filename}
                 will be derived.}
}
\value{
personography A data frame that holds the names and tcp numbers of the
                      people extracted from the \code{g} object.
}
\description{
Exports a personography from graph data.  The data is both returned
and written to a file with \code{filename}.
}
\examples{
personography = exportPersonography(g=g, filename="persons.csv", dnet=dnet)
}

