% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractData.R
\name{extractData}
\alias{extractData}
\title{A function that allows a user to extract data from
the \code{edges} of a \code{docNetwork} object and
saves it to a file.}
\usage{
extractData(dnet, file = "", append = F, sub = F, comms = NULL,
  by = 0)
}
\arguments{
\item{dnet}{The \code{docNetwork} object to modify.}

\item{file}{The full path to the file to write the
data to.}

\item{append}{A logical that indicates whether the
data should be added to an existing
file or not.  Default is False.}

\item{sub}{A logical that indicates whether the
desired data is a subset of the overall
network.}

\item{comms}{The community data from a \code{docNetwork}
object to be used in subsetting the data.}

\item{by}{The community number to use to subset the
                data with.}
}
\value{
df    A subset of \code{dnet}'s \code{index} for all nodes.

data  The vector of TCP numbers for a subset of the
              network's edges.
}
\description{
A function that allows a user to extract data from
the \code{edges} of a \code{docNetwork} object and
saves it to a file.
}
\examples{
df = extractData(dnet, file="/path/to/file.csv")
tcps = extractData(dnet, file="/path/to/file.csv", sub=T, comms=dnet[at]communities, by=2)
dnet = extractData(dnet, "/path/to/file.csv", append=T)
}

