% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEdges.R
\name{extractEdges}
\alias{extractEdges}
\title{Gets network data from EEBO-TCP index}
\usage{
extractEdges(dnet)
}
\description{
A function that extracts edges from an index.
 The edges are derived from the data included with
 this package. For whichever TCP files are included in your
 \code{docNetwork} index, this function finds the corresponding
 edges from the total data.
 
 @param dnet  A \code{docNetwork} object that contains an
              \code{index} that has \code{TCP} numbers.
 
 @return edges A \code{data.frame} that contains edge data
               in the form: \code{name}, \code{TCP}.
}

