% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htn.R
\docType{package}
\name{htn}
\alias{htn}
\alias{htn-package}
\title{htn: Model and analyze historical text networks}
\description{
The \code{htn} package buils on \code{tei2r} to perform social-network
analysis on bibliographic metadata, derived from the Text Creation
Partnership collection.
}
\section{What it does}{

The basic purpose of the \code{htn} package is to convert bibliographic
metadata to networks of co-publication. It comes with a dataset, called
\code{tcpEdges}, that includes the names of printers, booksellers,
and authors that each document is attributed to. It then converts that
data into the form of a social-network graph using the \code{igraph}
package.

The main functionality of the package is to identify networks and then
parse those networks into communities, allowing researchers to identify
subsets of the EEBO-TCP corpus based on co-publication data. These
subsets can then be downloaded, imported into R, and topic modeled using 
\code{tei2r}, or analyzed using \code{empson}, which has a variety of 
functions for text analysis.

For a tutorial that walks through the basics, check out the introductory 
vignette, "Introduction: Getting Started with htn."

\strong{Basic operations:}
\itemize{
  \item \code{\link{buildNetwork}}: Convert a \code{tei2r::docList} into
        a social network graph.
  \item \code{\link{communitySubnetwork}}: Identify subsets of the TCP
        collection that correspond to different clusters of co-publication.
  \item \code{\link{drawGraph}}: Visualize historical text networks.
}

\strong{Sample Data:}
\itemize{
  \item \code{\link{tcpEdges}}: A dataframe with source-target data over 
  the entire EEBO-TCP Phase I collection.
                              
  \item \code{\link{tcpPersons}}: A dataframe with name and role identifiers
  for each person included in the \code{tcpEdges} edge list
}
}
\author{
Authors:
Michael Gavin and Travis Mullen
}

