% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importEdges.R
\name{importEdges}
\alias{importEdges}
\title{Import edge data into a \code{docNetwork} object from a csv
file.  This is most useful for adding attributes to exported
edges or for adding new nodes.  It can also be used to import
a new set of network data by setting \code{append=F}.}
\usage{
importEdges(dnet, file, append = T, delim = ";")
}
\arguments{
\item{dnet}{The \code{docNetwork} object to modify.}

\item{file}{The full path to the csv file that contains
               the new data. Should be in the format: name; tcp--or other id number.}

\item{append}{A boolean that sets whether the data should be
               added to the \code{docNetwork} object's \code{edges}
               or if it should replace the existing--or non-existent-
               data. Default is True.}

\item{delim}{A character string that shows how the names in your
               data are separated.}
}
\value{
dnet  The modified \code{docNetwork} object.
}
\description{
Import edge data into a \code{docNetwork} object from a csv
file.  This is most useful for adding attributes to exported
edges or for adding new nodes.  It can also be used to import
a new set of network data by setting \code{append=F}.
}
\examples{
dnet = importEdges(dnet, file="/path/to/edge/data.csv", delim=',')
dnet = importEdges(dnet, file="/path/to/edge/data.csv")
dnet = importEdges(dnet, file="/path/to/edge/data.csv", append=F)
}

