% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importPersonography.R
\name{importPersonography}
\alias{importPersonography}
\title{Import a personography into a \code{docNetwork} object.
File must be in the format: ID, Role}
\usage{
importPersonography(dnet, path = "")
}
\arguments{
\item{dnet}{a \code{docNetwork} object to add the data to.}

\item{path}{A string that gives the full path to the csv
              formatted data file.}
}
\value{
dnet The modified dnet object.
}
\description{
Import a personography into a \code{docNetwork} object.
File must be in the format: ID, Role
}
\examples{
dnet = importPersonography(dnet, "path/to/file/data.csv")
}

