% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints.R
\docType{data}
\name{imprints}
\alias{imprints}
\title{Data - Imprints with Attributions}
\format{A dataframe with 64,887 observations and 5 variables.}
\usage{
imprints
}
\description{
A dataframe containing publication metadata for the EEBO-TCP
collection.
}
\examples{
# Load the collection, then setup on your computer
data(imprints)
View(imprints)
}
\keyword{datasets}

