% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viewCommunities.R
\name{viewCommunities}
\alias{viewCommunities}
\title{View the community data associated with the
graph stored in a \code{docNetwork} object.}
\usage{
viewCommunities(dnet)
}
\arguments{
\item{dnet}{The \code{docNetwork} object that
             holds the community data.}
}
\description{
View the community data associated with the
graph stored in a \code{docNetwork} object.
}

