% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{This function constructs a \code{docNetwork} object from
a \code{docList} object.  It will extract the edges and
persons, build the graph, and detect the communities
within that graph.  The default type of community
detection is \code{walktrap}.}
\usage{
buildNetwork(dl, edges = NULL, persons = NULL)
}
\arguments{
\item{dl}{A \code{docList} object that contains the \code{index}
             and other meta-data for the corpus that is to be
             networked.}

\item{edges}{A \code{data.frame} or string that holds the full path to
             a csv file with edge data in it. Must be in the form:
             \code{name}, \code{TCP} or other document identifier.
             Optional.}

\item{persons}{A \code{data.frame} or string that holds the full path to
               a csv file with persons data in it. Must be in the form:
               \code{ID}, \code{Role}. Optional.}
}
\description{
This function constructs a \code{docNetwork} object from
a \code{docList} object.  It will extract the edges and
persons, build the graph, and detect the communities
within that graph.  The default type of community
detection is \code{walktrap}.
}
\section{How to use it}{

For each of the parameters (other than the docList), the user can either
provide a data.frame, a character string that connects to a .csv file,
or let the default function extract the names.
}
\examples{
dnet = buildNetwork(dl)
dnet = buildNetwork(dl, edges="path/to/edges.csv", persons="path/to/persons.csv")
}

