% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectCommunities.R
\name{detectCommunities}
\alias{detectCommunities}
\title{Find communities in a graph}
\usage{
detectCommunities(dnet, type = "walktrap", view = T)
}
\arguments{
\item{dnet}{A document network}

\item{type}{The algorithm to be used: options include "walktrap" and "betweenness".}

\item{view}{A logical that indicates whether you want to view the communities
            after detecting them or not. Default is True.}
}
\value{
comms The \code{communities} object that holds the communities that
              were detected in \code{g}
}
\description{
Find communities in a graph
}
\examples{
communities = detectCommunities(dnet, type="edge.betweenness")
communities = detectCommunities(dnet, view=F)
comms = detectCommunities(dnet)
}

