% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEdges.R
\name{extractEdges}
\alias{extractEdges}
\title{A function that extracts edges from a dnet object's index.
 The edges are derived from the sample data included with
 this package and based upon the \code{TCP} numbers included
 in the \code{docNetwork}'s \code{index}.
 
 @param dnet  A \code{docNetwork} object that contains an
              \code{index} that has \code{TCP} numbers.
 
 @return edges A \code{data.frame} that contains edge data
               in the form: \code{name}, \code{TCP}.}
\usage{
extractEdges(dnet)
}
\description{
A function that extracts edges from a dnet object's index.
 The edges are derived from the sample data included with
 this package and based upon the \code{TCP} numbers included
 in the \code{docNetwork}'s \code{index}.
 
 @param dnet  A \code{docNetwork} object that contains an
              \code{index} that has \code{TCP} numbers.
 
 @return edges A \code{data.frame} that contains edge data
               in the form: \code{name}, \code{TCP}.
}

